+++ src/main/java/org/patternfly/layout/bullseye/Bullseye.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.patternfly.layout.gallery;
+package org.patternfly.layout.bullseye;
 
 import org.jboss.elemento.HTMLContainerBuilder;
 import org.patternfly.layout.BaseLayout;
@@ -21,32 +21,42 @@ import org.patternfly.layout.BaseLayout;
 import elemental2.dom.HTMLElement;
 
 import static org.jboss.elemento.Elements.div;
-import static org.patternfly.style.Classes.gallery;
-import static org.patternfly.style.Classes.item;
+import static org.patternfly.style.Classes.bullseye;
 import static org.patternfly.style.Classes.layout;
 
-public class GalleryItem extends BaseLayout<HTMLElement, GalleryItem> {
+/**
+ * The bullseye layout centers content, both vertically and horizontally within a container.
+ *
+ * @see <a href="https://www.patternfly.org/layouts/bullseye">https://www.patternfly.org/layouts/bullseye</a>
+ */
+public class Bullseye extends BaseLayout<HTMLElement, Bullseye> {
 
     // ------------------------------------------------------ factory
 
-    public static GalleryItem galleryItem() {
-        return new GalleryItem(div());
+    public static Bullseye bullseye() {
+        return new Bullseye(div());
     }
 
-    public static <E extends HTMLElement> GalleryItem galleryItem(HTMLContainerBuilder<E> builder) {
-        return new GalleryItem(builder);
+    public static <E extends HTMLElement> Bullseye bullseye(HTMLContainerBuilder<E> builder) {
+        return new Bullseye(builder);
     }
 
     // ------------------------------------------------------ instance
 
-    <E extends HTMLElement> GalleryItem(HTMLContainerBuilder<E> builder) {
-        super(builder.css(layout(gallery, item)).element());
+    <E extends HTMLElement> Bullseye(HTMLContainerBuilder<E> builder) {
+        super(builder.css(layout(bullseye)).element());
+    }
+
+    // ------------------------------------------------------ add
+
+    public Bullseye addItem(BullseyeItem item) {
+        return add(item);
     }
 
     // ------------------------------------------------------ builder
 
     @Override
-    public GalleryItem that() {
+    public Bullseye that() {
         return this;
     }
 }